services/arch-engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pyyaml==6.0.1
python-dotenv==1.0.0
gitpython==3.1.41
docker==7.0.0
jinja2==3.1.3
